+++ Project24/Pages/Home/Management/ConfigPanel.cshtml.cs
@@ -18,7 +18,7 @@ using Project24.App;
 using Project24.App.Services;
 using Project24.App.Utils;
 using Project24.Data;
-using Project24.SerializerContext;
+using Project24.Serializer;
 
 namespace Project24.Pages.Home.Management
 {